Bitti sanırım

+++ code/controllers/subsystem/storyteller.dm
@@ -5,6 +5,8 @@ SUBSYSTEM_DEF(storyteller)
 
 	var/datum/storyteller/current_storyteller
 
+	var/datum/storyteller/next_storyteller
+
 	var/list/storyteller_prototypes = list()
 	var/list/storyteller_namelist = list()
 
@@ -28,18 +30,38 @@ SUBSYSTEM_DEF(storyteller)
 		.[storyteller.name] = storyteller
 
 /datum/controller/subsystem/storyteller/proc/load_storyteller(list/dynamic_config)
-	var/selected_storyteller_name = trim(file2text("data/next_round_storyteller.txt"))
+	var/default_storyteller = CONFIG_GET(string/default_storyteller)
 
-	var/datum/storyteller/selected_storyteller
+	var/list/storyteller_data = load_storyteller_data() // Load storyteller from data/next_round_storyteller.json
+	var/selected_storyteller_name = storyteller_data["name"]
+	var/forced = storyteller_data["forced"]
 
-	if(!selected_storyteller_name || !storyteller_namelist[selected_storyteller_name])
+	var/select_random = FALSE
+
+	if((!CONFIG_GET(flag/storyteller_votable) && !forced) || !storyteller_namelist[selected_storyteller_name]) //If (its not chosen by admin and storyteller_votable flag is false) or its not a valid storyteller, make it randomised
+		if(!storyteller_namelist[selected_storyteller_name])
+			stack_trace("Storyteller: [selected_storyteller_name] is not a valid storyteller!")
+		select_random = TRUE
+
+	if(default_storyteller != "Random" && storyteller_namelist[default_storyteller] && select_random) //If default_storyteller is not "Random" and its a valid storyteller and select_random is true, use it
+		selected_storyteller_name = default_storyteller
+		select_random = FALSE
+
+	if(select_random)
 		var/list/storyteller_entries = get_valid_storytellers(weighted = TRUE)
 		selected_storyteller_name = pick_weight(storyteller_entries)
 
-	selected_storyteller = storyteller_namelist[selected_storyteller_name]
+	var/datum/storyteller/selected_storyteller = storyteller_namelist[selected_storyteller_name]
 	var/list/config = SSdynamic.get_config()
 	current_storyteller = new selected_storyteller.type(config)
 	log_game("Storyteller loaded: [current_storyteller.name]")
+
+/datum/controller/subsystem/storyteller/proc/load_storyteller_data()
+	var/json_file = file("data/next_round_storyteller.json")
+	if(!fexists(json_file))
+		return list()
+
+	return json_decode(file2text(json_file))
 
 /datum/controller/subsystem/storyteller/proc/get_valid_storytellers(weighted = FALSE)
 	var/filter_threshold = 0
@@ -52,12 +74,6 @@ SUBSYSTEM_DEF(storyteller)
 	var/list/storytellers = shuffle(storyteller_prototypes)
 
 	for(var/datum/storyteller/storyteller in storytellers)
-		if(storyteller.always_votable)
-			if(weighted)
-				valid_storytellers[storyteller.name] = storyteller.weight
-			else
-				valid_storytellers += storyteller.name
-			continue
 		if(storyteller.restricted)
 			continue
 		if(storyteller.population_min > 0 && filter_threshold < storyteller.population_min)
@@ -77,12 +93,42 @@ SUBSYSTEM_DEF(storyteller)
 	if(storyboy)
 		return storyboy.desc
 
-/datum/controller/subsystem/storyteller/proc/admin_set_storyteller(datum/storyteller/storyteller_type)
+/datum/controller/subsystem/storyteller/proc/set_storyteller(storyteller_name, for_current_round = FALSE, forced = FALSE)
+	. = TRUE
 	var/list/config = SSdynamic.get_config()
-	var/datum/storyteller/old_storyteller = current_storyteller
-	current_storyteller = new storyteller_type(config)
-	SSdynamic.set_tier(SSdynamic.current_tier.type, SSticker.totalPlayersReady) //Reload the dynamic tier
-	qdel(old_storyteller)
+	var/datum/storyteller/storyteller_prototype = storyteller_namelist[storyteller_name]
+	if(isnull(storyteller_prototype))
+		stack_trace("Storyteller: [storyteller_name] is not a valid storyteller!")
+		return FALSE
+	if(for_current_round) //If the proc called for the current round, change it immediately
+		var/datum/storyteller/old_storyteller = current_storyteller
+		current_storyteller = new storyteller_prototype.type(config)
+		if(!isnull(SSdynamic.current_tier))
+			SSdynamic.set_tier(SSdynamic.current_tier.type, SSticker.totalPlayersReady) //Reload the dynamic tier
+		qdel(old_storyteller)
+		message_admins("[key_name_admin(usr)] changed storyteller to [storyteller_name].")
+		log_admin("[key_name_admin(usr)] changed storyteller to [storyteller_name].")
+	else //If the proc called for the next round, save it to json
+		var/datum/storyteller/old_storyteller = next_storyteller
+		next_storyteller = new storyteller_prototype.type //No need to use config, it will handle in the next round
+		var/list/next_storyteller_data = list(
+			"name" = storyteller_name,
+			"forced" = forced
+		)
+		var/next_storyteller_file = "data/next_round_storyteller.json"
+		if (fexists(next_storyteller_file))
+			fdel(next_storyteller_file)
+		WRITE_FILE(file(next_storyteller_file), json_encode(next_storyteller_data, JSON_PRETTY_PRINT))
+		qdel(old_storyteller)
+		message_admins("[key_name_admin(usr)] set next round's storyteller to [storyteller_name].")
+		log_admin("[key_name_admin(usr)] set next round's storyteller to [storyteller_name].")
+
+/datum/controller/subsystem/storyteller/vv_edit_var(var_name, var_value)
+	if(var_name == NAMEOF(src, current_storyteller))
+		return FALSE
+	if(var_name == NAMEOF(src, next_storyteller))
+		return FALSE
+	return ..()
 
 /datum/controller/subsystem/storyteller/Recover()
 	current_storyteller = SSstoryteller.current_storyteller
